Split indexer into index-updater and index-query-engine
