+++ packages/runtime-common/index.ts
@@ -44,7 +44,9 @@ import { Loader } from './loader';
 export * from './constants';
 export * from './queue';
 export * from './expression';
-export * from './indexer';
+export * from './index-query-engine';
+export * from './index-updater';
+export * from './index-structure';
 export * from './db';
 export * from './worker';
 export * from './stream';
